==> ExAC/13.Process_domains_not_in_training/1.Find_processed_domains_not_in_pipeline.ipynb <==
import pandas as pd
import numpy as np
import pickle
from collections import defaultdict

curr_dir = !pwd
#Read the InteracDome propensity
with open(curr_dir[0]+"/../7.InteracDome_targetVariable/domains_ligands_propensity_dict.pik", 'rb') as handle:
        binding_props_dict = pickle.load(handle)
#Read the list of processed domains not in pipeline (<10 instances)
ls_domains_less_than_10 = pd.read_csv("../6.Ext_features_to_dicts/ext_features_dicts/less_than_10_domains.txt", header=None)

less_than_10_domains_list = []
for index,row in ls_domains_less_than_10.iterrows():
    less_than_10_domains_list.append(row[0][:-1])

#Find which processed domains have InteracDome scores
interacdome_domains_set = set(binding_props_dict.keys())
processed_domains_set = set(less_than_10_domains_list)
intersect_domains = list(processed_domains_set.intersection(interacdome_domains_set))

len(intersect_domains)

#Save to file
with open(curr_dir[0]+"/processed_domains_with_labels_not_in_pipeline.pik", 'wb') as handle:
    pickle.dump(intersect_domains, handle, protocol=pickle.HIGHEST_PROTOCOL)
