==> ExAC/13.Process_domains_not_in_training/3.run_hhalign.ipynb <==
## Running hhalign from the HHsuite
Goal: create hhm files and run hhalign between them for all the relevant domains.

First, clone the HHsuite github for HHalign version 3.0.3 (14-07-2017) from: https://github.com/soedinglab/hh-suite according to their README

Than, install (make + sudo make install) into: /usr/local/hhsuite

After installation, create inside the cloned github directory "databases" directory and download "pfamA_31.0.tgz" from: http://wwwuser.gwdg.de/~compbiol/data/hhsuite/databases/hhsuite_dbs/ (and unzip there). This is the database of hhm (not hmm!) files for pfam-v31, which is the prefered format for using HHsuite tools.

import subprocess
import pandas as pd
import pickle
import os.path
from collections import defaultdict

curr_dir = !pwd
pfam_version = "31"

sim_pairs_df_filtered = pd.read_csv("processed_domains_pairs_for_hhalign.csv", sep=',', index_col=0)
with open(curr_dir[0]+"/../2.parse_Pfam/v"+pfam_version+"/domain_to_pfam_acc_dict.pik", 'rb') as handle:
    domain_to_pfam_acc_dict = pickle.load(handle)    
    
hhsuite_hhm_database_path = "../../../HHsuite/databases/pfam-v"+pfam_version+"/"
hhsuite_hhm_database_filename = "pfam_hhm.ffdata"
hhsuite_hhm_database_idx = "pfam_hhm.ffindex"

def create_domain_hhm_file(domain_pfam_id, start_pos, length):
    
    outfilename = "../10.Prediction/domains_similarity/pfam-v"+pfam_version+"/hhm_files/"+domain_pfam_id+".hhm"
    creating_hhm_cmd = "dd if=../../../HHsuite/databases/pfam-v31/pfam_hhm.ffdata of="+outfilename+" bs=1 count="+length+" skip="+start_pos
    not_important_prints = !$creating_hhm_cmd

def process_query(query, res_dict, direction):
    
    
    for line in query:
        
        #Get the probability of the query and the template to be homologus
        if (line.find("Probab=") >= 0):
            end_of_line = line[line.find("Probab=")+7:]
            prob = end_of_line[:end_of_line.find(" ")]
            res_dict["prob"+str(direction)].append(prob)
        
        #Get the alignment p-value
        if (line.find("E-value=") >= 0):
            end_of_line = line[line.find("E-value=")+8:]
            pval = end_of_line[:end_of_line.find(" ")]
            res_dict["pval"+str(direction)].append(pval)
            
        #Get the alignment raw score
        if (line.find("Score=") >= 0):
            end_of_line = line[line.find("Score=")+6:]
            score = end_of_line[:end_of_line.find(" ")]
            res_dict["score"+str(direction)].append(score)
        
        #Get the number of aligned columns
        if (line.find("Aligned_cols=") >= 0):
            end_of_line = line[line.find("Aligned_cols=")+13:]
            aligned = end_of_line[:end_of_line.find(" ")]
            res_dict["aligned_cols"+str(direction)].append(aligned)
        
        #Get the percentage of aligned residue pairs of the query and the template master sequences that are identical.
        if (line.find("Identities=") >= 0):
            end_of_line = line[line.find("Identities=")+11:]
            ident = end_of_line[:end_of_line.find("%")]
            res_dict["ident_perc"+str(direction)].append(ident)

%%time
res_dict = defaultdict(list)

header = "/usr/local/hhsuite/bin/hhalign -i "
ext = ".hhm"
flag = " -t "
hhm_path = "../10.Prediction/domains_similarity/pfam-v"+pfam_version+"/hhm_files/"
flags = " -hide_cons -hide_pred -hide_dssp -o hhalign_out"

for index, pair in sim_pairs_df_filtered.iterrows():
    #Get the domains names
    dom1 =  pair["sim_dom1"]
    dom2 =  pair["sim_dom2"]
    
    #Get the domains PFAM ids
    dom1_pfam_id = domain_to_pfam_acc_dict[dom1]
    dom2_pfam_id = domain_to_pfam_acc_dict[dom2]
    
    outfilename1 = "../10.Prediction/domains_similarity/pfam-v"+pfam_version+"/hhm_files/"+dom1_pfam_id+".hhm"
    if (os.path.isfile(outfilename1) == False):
        #Get domain 1 indices for fast access
        dom1_idx_cmd = "less "+hhsuite_hhm_database_path+hhsuite_hhm_database_idx+" | grep "+dom1_pfam_id
        dom1_idx_line = !$dom1_idx_cmd
        dom1_idx_line = dom1_idx_line[0].split("\t")
        dom1_start_pos = dom1_idx_line[1]
        dom1_len = dom1_idx_line[2]
        #Create the .hhm (special HHsuite format for HMM profiles) for domain 1
        create_domain_hhm_file(dom1_pfam_id,dom1_start_pos,dom1_len)
    
    outfilename2 = "../10.Prediction/domains_similarity/pfam-v"+pfam_version+"/hhm_files/"+dom2_pfam_id+".hhm"
    if (os.path.isfile(outfilename2) == False):
        #Get domain 2 indices for fast access
        dom2_idx_cmd = "less "+hhsuite_hhm_database_path+hhsuite_hhm_database_idx+" | grep "+dom2_pfam_id
        dom2_idx_line = !$dom2_idx_cmd
        dom2_idx_line = dom2_idx_line[0].split("\t")
        dom2_start_pos = dom2_idx_line[1]
        dom2_len = dom2_idx_line[2]
        #Create the .hhm (special HHsuite format for HMM profiles) for domain 2
        create_domain_hhm_file(dom2_pfam_id,dom2_start_pos,dom2_len)
    
    #Running hhalign direction #1
    script = header+hhm_path+dom1_pfam_id+ext+flag+hhm_path+dom2_pfam_id+ext+flags
    query = !$script
    process_query(query, res_dict, 1)
    #Running hhalign direction #2
    script = header+hhm_path+dom2_pfam_id+ext+flag+hhm_path+dom1_pfam_id+ext+flags
    query = !$script
    process_query(query, res_dict, 2)
    
    if (index % 100 == 0):
        print index

#Add infomration to the domains pairs table
sim_pairs_df_filtered["prob1"] = res_dict["prob1"]
sim_pairs_df_filtered["pval1"] = res_dict["pval1"]
sim_pairs_df_filtered["score1"] = res_dict["score1"]
sim_pairs_df_filtered["aligned_cols1"] = res_dict["aligned_cols1"]
sim_pairs_df_filtered["ident_perc1"] = res_dict["ident_perc1"]

sim_pairs_df_filtered["prob2"] = res_dict["prob2"]
sim_pairs_df_filtered["pval2"] = res_dict["pval2"]
sim_pairs_df_filtered["score2"] = res_dict["score2"]
sim_pairs_df_filtered["aligned_cols2"] = res_dict["aligned_cols2"]
sim_pairs_df_filtered["ident_perc2"] = res_dict["ident_perc2"]

#Save the file with all the results
sim_pairs_df_filtered.to_csv("processed_domains_hhlign_scores.csv", sep='\t')

sim_pairs_df_filtered
