==> ExAC/13.Process_domains_not_in_training/6.Filter_by_domains-mediode-final_domains_list.ipynb <==
## Leaving only mediode of each similarity cluster

import pandas as pd
import numpy as np
import pickle
from collections import defaultdict

from IPython.core.display import HTML
HTML("<style>.container { width:100% !important; }</style>")

### Reading the input dataset

curr_dir = !pwd
pfam_version = "31"

#similarity clusters
with open(curr_dir[0]+"/within_processed_sim_clusters/domain_to_cluster_dict.pik", 'rb') as handle:
    domain_to_cluster_dict = pickle.load(handle)
with open(curr_dir[0]+"/within_processed_sim_clusters/similarity_clusters_dict.pik", 'rb') as handle:
    similarity_clusters_dict = pickle.load(handle)
print "clusters #: "+str(len(similarity_clusters_dict.keys()))
print "domains in clusters #: "+str(len(domain_to_cluster_dict.keys()))

### Creating the final processed domains list

#Reading the non-similar domains list
with open(curr_dir[0]+"/processed_domains_non_similar_at_all.pik", 'rb') as handle:
    non_similar_domains_list = pickle.load(handle)

### Adding the clusters mediodes to the list

#Finding which domains we need to remove
domains_mediodes = []
for cluster_num in range(len(similarity_clusters_dict.keys())):
    domains_mediodes.append(similarity_clusters_dict[cluster_num]["mediode"][0]) #If there's more than one mediode, choosing one randomly

len(domains_mediodes)

final_processed_domains_list = non_similar_domains_list + domains_mediodes
len(final_processed_domains_list)

#Verifying no duplicates
len(set(final_processed_domains_list))

#Saving to file
with open(curr_dir[0]+"/processed_domains_not_in_pipeline_final_list.pik", 'wb') as handle:
    pickle.dump(final_processed_domains_list, handle, protocol=pickle.HIGHEST_PROTOCOL)
